refactor: remove login methods to components
